攻撃モーション用のファイルパスを追加，skinned_mesh.hの警告を消した

+++ Chi/Chi/skinned_mesh.h
@@ -177,7 +177,8 @@ public:
 	
 	void setLoopFlg(const bool _is_loop) { loop_flg = _is_loop; }
 	void setStopAnimation(const bool _is_stop) { stop_animation = _is_stop; }
-	void setStoptimer(const int _stop_timer) { stop_time = _stop_timer; }
+	// void setStoptimer(const int _stop_timer) { stop_time = _stop_timer; }
+	void setStoptimer(const int _stop_timer) { stop_time = static_cast<float>( _stop_timer ); }
 
 private:
 	//template <class Archive>
